+++ src/uisupport/nickview.cpp
@@ -25,8 +25,6 @@
 #include <QScrollBar>
 #include <QDebug>
 #include <QMenu>
-#include <QTouchEvent>
-#include <QScrollBar>
 
 #include "buffermodel.h"
 #include "client.h"
@@ -38,7 +36,7 @@
 #include "types.h"
 
 NickView::NickView(QWidget *parent)
-    : QTreeView(parent)
+    : TreeViewTouch(parent)
 {
     setIndentation(10);
     header()->hide();
@@ -74,7 +72,6 @@ void NickView::init()
 
     connect(selectionModel(), SIGNAL(currentChanged(QModelIndex, QModelIndex)), SIGNAL(selectionUpdated()));
     connect(selectionModel(), SIGNAL(selectionChanged(QItemSelection, QItemSelection)), SIGNAL(selectionUpdated()));
-	setAttribute(Qt::WA_AcceptTouchEvents);
 }
 
 
@@ -83,14 +80,14 @@ void NickView::setModel(QAbstractItemModel *model_)
     if (model())
         disconnect(model(), 0, this, 0);
 
-    QTreeView::setModel(model_);
+    TreeViewTouch::setModel(model_);
     init();
 }
 
 
 void NickView::rowsInserted(const QModelIndex &parent, int start, int end)
 {
-    QTreeView::rowsInserted(parent, start, end);
+    TreeViewTouch::rowsInserted(parent, start, end);
     if (model()->data(parent, NetworkModel::ItemTypeRole) == NetworkModel::UserCategoryItemType && !isExpanded(parent)) {
         unanimatedExpandAll();
     }
@@ -107,7 +104,7 @@ void NickView::setRootIndex(const QModelIndex &index)
 
 QModelIndexList NickView::selectedIndexes() const
 {
-    QModelIndexList indexList = QTreeView::selectedIndexes();
+    QModelIndexList indexList = TreeViewTouch::selectedIndexes();
 
     // make sure the item we clicked on is first
     if (indexList.contains(currentIndex())) {
@@ -152,40 +149,4 @@ void NickView::startQuery(const QModelIndex &index)
         return;
 
     Client::bufferModel()->switchToOrStartQuery(networkId, ircUser->nick());
-}
-
-bool NickView::event(QEvent *event) {
-	if (event->type() == QEvent::TouchBegin && _lastTouchStart < QDateTime::currentMSecsSinceEpoch() - 1000) { //(slow) double tab = normal behaviour = select multiple. 1000 ok?
-		_touchScrollInProgress = true;
-		_lastTouchStart = QDateTime::currentMSecsSinceEpoch();
-		setVerticalScrollMode(QAbstractItemView::ScrollPerPixel);
-		return true;
-	}
-
-	if (event->type() == QEvent::TouchUpdate && _touchScrollInProgress) {
-		QTouchEvent::TouchPoint p = ((QTouchEvent*)event)->touchPoints().at(0);
-		verticalScrollBar()->setValue(verticalScrollBar()->value() - (p.pos().y() - p.lastPos().y()));
-		return true;
-	}
-
-#if QT_VERSION >= 0x050000
-	if (event->type() == QEvent::TouchEnd || event->type() == QEvent::TouchCancel) {
-#else
-    if (event->type() == QEvent::TouchEnd) {
-#endif
-		_touchScrollInProgress = false;
-		return true;
-	}
-
-	return QTreeView::event(event);
-}
-
-void NickView::mousePressEvent(QMouseEvent * event) {
-	if (!_touchScrollInProgress)
-		QTreeView::mousePressEvent(event);
-}
-
-void NickView::mouseMoveEvent(QMouseEvent * event) {
-	if (!_touchScrollInProgress)
-		QTreeView::mouseMoveEvent(event);
 }
