+++ src/uisupport/nickview.h
@@ -24,8 +24,9 @@
 #include <QTreeView>
 
 #include "bufferinfo.h"
+#include "treeviewtouch.h"
 
-class NickView : public QTreeView
+class NickView : public TreeViewTouch
 {
     Q_OBJECT
 
@@ -37,9 +38,6 @@ protected:
 
     //! This reimplementation ensures that the current index is first in list
     virtual QModelIndexList selectedIndexes() const;
-	virtual bool event(QEvent *event);
-	virtual void mouseMoveEvent(QMouseEvent *event);
-	virtual void mousePressEvent(QMouseEvent *event);
 
     void unanimatedExpandAll();
 
@@ -55,8 +53,6 @@ signals:
 
 private:
     friend class NickListWidget; // needs selectedIndexes()
-	qint64 _lastTouchStart = 0;
-	bool _touchScrollInProgress = false;
 };
